// src/api/openapi/models/CloseTripOfferReason.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const CloseTripOfferReason = {
    DEMANDER_GENERAL: 'DEMANDER_GENERAL',
    DEMANDER_UNPAID: 'DEMANDER_UNPAID',
    TRANSPORTER_GENERAL: 'TRANSPORTER_GENERAL',
    GENERAL: 'GENERAL'
} as const;
export type CloseTripOfferReason = typeof CloseTripOfferReason[keyof typeof CloseTripOfferReason];


export function instanceOfCloseTripOfferReason(value: any): boolean {
    for (const key in CloseTripOfferReason) {
        if (Object.prototype.hasOwnProperty.call(CloseTripOfferReason, key)) {
            if (CloseTripOfferReason[key as keyof typeof CloseTripOfferReason] === value) {
                return true;
            }
        }
    }
    return false;
}

export function CloseTripOfferReasonFromJSON(json: any): CloseTripOfferReason {
    return CloseTripOfferReasonFromJSONTyped(json, false);
}

export function CloseTripOfferReasonFromJSONTyped(json: any, ignoreDiscriminator: boolean): CloseTripOfferReason {
    return json as CloseTripOfferReason;
}

export function CloseTripOfferReasonToJSON(value?: CloseTripOfferReason | null): any {
    return value as any;
}

export function CloseTripOfferReasonToJSONTyped(value: any, ignoreDiscriminator: boolean): CloseTripOfferReason {
    return value as CloseTripOfferReason;
}