// src/api/openapi/models/TripOfferState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TripOfferState = {
    CREATED: 'CREATED',
    HAS_OFFERS: 'HAS_OFFERS',
    ACCEPTED_TRANSPORTER_PAY_DEPOSIT: 'ACCEPTED_TRANSPORTER_PAY_DEPOSIT',
    ACCEPTED_TRANSPORTER_PAY_FULL: 'ACCEPTED_TRANSPORTER_PAY_FULL',
    PAYED_DEPOSIT: 'PAYED_DEPOSIT',
    PAYED_FULL: 'PAYED_FULL',
    STARTED: 'STARTED',
    FINISHED: 'FINISHED',
    CLOSED: 'CLOSED'
} as const;
export type TripOfferState = typeof TripOfferState[keyof typeof TripOfferState];


export function instanceOfTripOfferState(value: any): boolean {
    for (const key in TripOfferState) {
        if (Object.prototype.hasOwnProperty.call(TripOfferState, key)) {
            if (TripOfferState[key as keyof typeof TripOfferState] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TripOfferStateFromJSON(json: any): TripOfferState {
    return TripOfferStateFromJSONTyped(json, false);
}

export function TripOfferStateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TripOfferState {
    return json as TripOfferState;
}

export function TripOfferStateToJSON(value?: TripOfferState | null): any {
    return value as any;
}

export function TripOfferStateToJSONTyped(value: any, ignoreDiscriminator: boolean): TripOfferState {
    return value as TripOfferState;
}